backend/python_engine/requirements.txt
# Core dependencies for deployment
fastapi==0.105.0
uvicorn[standard]==0.24.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.25.2

# Data processing
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Caching
redis==5.0.1

# AI/ML
openai==1.10.0
anthropic==0.34.0

# LangChain ecosystem
langchain==0.1.20
langgraph==0.2.16
langchain-chroma==0.1.2
langchain-openai==0.1.25
langchain-anthropic==0.1.20
langchain-core==0.1.52
langchain-text-splitters==0.1.0

# DSPy
dspy-ai==2.5.5

# Streaming
sse-starlette==1.8.2

# HTTP requests
requests==2.31.0

# Additional dependencies
python-multipart==0.0.6
aiofiles==23.2.0

# Vector database
chromadb==0.4.24
faiss-cpu==1.7.4

# Testing (minimal for production)
pytest==7.4.4
pytest-asyncio==0.21.1